+++ src/util.rs
@@ -16,7 +16,7 @@ use crate::{
 };
 
 /// Generates a `RSA` private key.
-pub fn generate_rsa_key() -> Result<Rsa<Private>, Error> {
+pub(crate) fn generate_rsa_key() -> Result<Rsa<Private>, Error> {
     Ok(Rsa::generate(KEY_WIDTH)?)
 }
 
@@ -29,7 +29,16 @@ pub fn generate_rsa_keypair() -> Result<(Rsa<Private>, Rsa<Public>), Error> {
     ))
 }
 
-/// Builds a json web key `JWK` (RFC7517).
+/// Builds a json web key `JWK` (RFC7517) for a RSA private key.
+/// # Example
+/// ```rust
+/// use acme_rs::util::jwk;
+/// use openssl::Rsa;
+///
+/// let priv_key = Rsa::generate(2048).expect("Error while receiving private key");
+///
+/// let jwk = jwk(priv_key).expect("Error while creating jwk");
+/// ```
 pub fn jwk(private_key: &Rsa<Private>) -> Result<serde_json::Value, Error> {
     let e = b64(&private_key.e().to_vec());
     let n = b64(&private_key.n().to_vec());
@@ -41,7 +50,30 @@ pub fn jwk(private_key: &Rsa<Private>) -> Result<serde_json::Value, Error> {
     }))
 }
 
-/// Constructs a json web signature `JWS` (RFC7515) in the flattened `JSON` form.
+/// Constructs a json web signature `JWS` (RFC7515) in the flattened `JSON` form for a specified
+/// payload. This involves signing the JWS with the RS256 algorithm.
+/// # Example
+/// ```rust
+/// use acme_rs::util::jws;
+/// use serde_json::json;
+/// use openssl::Rsa;
+///
+/// // get the private key, the jws header and the payload
+/// let priv_key = Rsa::generate(2048).expect("Error while receiving private key");
+/// let header = json!({
+///    "alg": "RS256",
+///    "nonce": "superRandom",
+/// });
+///
+/// let payload = json!({
+///    "termsOfServiceAgreed": true,
+///    "email": "foo@bar.de"
+/// });
+///
+/// // calculate the jws
+/// let jws = jws(payload, header, &priv_key).expect("Error while creating jws");
+///
+/// ```
 pub fn jws(
     payload: serde_json::Value,
     header: serde_json::Value,
@@ -74,13 +106,13 @@ pub fn jws(
 }
 
 /// Returns the `base64url` encoding of the input.
-pub fn b64(to_encode: &[u8]) -> String {
+pub(crate) fn b64(to_encode: &[u8]) -> String {
     encode_config(to_encode, base64::URL_SAFE_NO_PAD)
 }
 
 /// Extracts the payload and `replay-nonce` header field from a given http `Response`.
 #[inline]
-pub fn extract_payload_and_nonce<T>(response: Response) -> Result<(Nonce, T), Error>
+pub(crate) fn extract_payload_and_nonce<T>(response: Response) -> Result<(Nonce, T), Error>
 where
     T: DeserializeOwned,
 {
@@ -97,7 +129,7 @@ where
 /// Extracts the `location` and `replay-nonce` header field as well as
 /// the payload from a given http `Response`.
 #[inline]
-pub fn extract_payload_location_and_nonce<T>(
+pub(crate) fn extract_payload_location_and_nonce<T>(
     response: Response,
 ) -> Result<(String, Nonce, T), Error>
 where
@@ -142,7 +174,7 @@ pub fn save_certificates(certificate_chain: Certificate) -> Result<(), Error> {
     Ok(())
 }
 
-/// Saves an rsa keypair into two files `priv.pem` and `pub.pem`
+/// Saves an rsa keypair into two files `priv.pem` and `pub.pem`.
 pub fn save_keypair(keypair: &(Rsa<Private>, Rsa<Public>)) -> Result<(), Error> {
     let private_key = keypair.0.private_key_to_pem()?;
     let public_key = keypair.1.public_key_to_pem()?;
